Refactors config to allow checking if adapter is set
To address a concern of some of the adapters, we need the ability to
check whether the adapter has been set. By adding this method, we
provide the ability to conditionally set the adapter while maintaining
the exceptions that currently are raised if the adapter is changed (or
accessed without being set).

This also modifies the way config loads, namely that the defaults are
set in an initializer and we use attrs to configure the various
attributes.

It also changes the hostname lookup to use `Socket` instead. This helps
keep the library more OS agnostic versus the old shell-out method.

In addition, the specs were expanded to provide additional rigor around
how the config object works.