Adds ability to temporarily modify the local mode (#18)
Previously, changing the configuration for local_mode would result in
the setting changing for all threads. This caused issues when run in
sidekiq because there can be multiple threads performing work at the
same time.

Now, there is a new method (with_local_mode) which will execute a block
with a particular local mode. The value is stored on a thread local
variable so that it does not bleed into other executions.