Fixes `with_local_mode` for ruby < 2.6
In ruby 2.6, the thread variables get deleted if set to nil. However, in
older versions of ruby this appears to not be the case. This patch
changes the behavior so that we use the type of the value to determine
if we should use it or fall back to the instance variable. This works in
2.6 and 2.5